requirements.txt
absl-py==0.10
appdirs==1.4.4
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.11
cycler==0.11.0
Cython==0.29.28
dataclasses==0.7
decorator==4.4.2
ez-setup==0.9
flatbuffers==1.12.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.0
h5py==3.1.0
idna==3.3
imageio==2.14.1
importlib-metadata==4.8.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
MedPy==0.4.0
meson==0.55.1
mpi4py==3.1.3
networkx==2.5.1
numba==0.53.1
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.4
psutil==5.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytools==2021.2.9
PyWavelets==1.1.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.4.1
scikit-image==0.17.2
SimpleITK==2.1.1.2
six==1.15.0
tensorboard==2.5
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
tifffile==2020.9.3
typing-extensions==3.7.4
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
